+++ src/lib/main/grid/grid.component.ts
@@ -190,14 +190,15 @@ export class GridComponent extends RootComponent implements OnInit {
 			}));
 		});
 
+		this.using(listener.on('keydown', e => {
+			const result = ctrl.keyDown(e, 'grid');
+			if (result.indexOf('selection.view') >= 0) {
+				this.cd.markForCheck();
+				this.zone.run(noop);
+			}
+		}));
+
 		this.zone.runOutsideAngular(() => {
-			this.using(listener.on('keydown', e => {
-				const result = ctrl.keyDown(e, 'grid');
-				if (result.indexOf('selection.view') >= 0) {
-					this.cd.markForCheck();
-					this.zone.run(noop);
-				}
-			}));
 			this.using(listener.on('keyup', e => ctrl.keyUp(e, 'grid')));
 		});